+++ lang/pt/theme_stream.php
@@ -1,0 +1,163 @@
+<?php
+// This file is part of Moodle - https://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <https://www.gnu.org/licenses/>.
+
+/**
+ * Plugin strings are defined here.
+ *
+ * @package     theme_stream
+ * @category    string
+ * @copyright   2022 Hugo Ribeiro <ribeiro.hugo@gmail.com>
+ * @license     https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+defined('MOODLE_INTERNAL') || die();
+
+$string['pluginname'] = 'Stream';
+$string['choosereadme'] = 'Tema Stream';
+$string['hiddencourse'] = 'Este curso está oculto para estudantes.';
+$string['frontendby'] = 'Através do tema Stream';
+$string['generalsettings'] = 'Configurações gerais';
+$string['courseheaderimg'] = 'Imagem standard de cabeçalho';
+$string['courseheaderimg_desc'] = 'Imagem comum para ser usada nos cabeçalhos dos cursos caso não exista uma particular no contexto do curso';
+$string['footersettings'] = 'Conteúdo para rodapé';
+$string['socialmedia'] = 'Redes Sociais';
+$string['facebookurl'] = 'Endereço Facebook';
+$string['instagramurl'] = 'Endereço Instagram';
+$string['pinteresturl'] = 'Endereço Pinterest';
+$string['youtubeurl'] = 'Endereço Youtube';
+$string['linkedinurl'] = 'Endereço Linkedin';
+$string['twitterurl'] = 'Twitter URL';
+$string['region-topblock'] = 'Área topo da disciplina';
+$string['colours'] = 'Definição de cores';
+$string['primarycolour'] = 'Cor principal';
+$string['primarycolour_desc'] = 'Cor da marca';
+$string['secondarycolour'] = 'Cor secundária';
+$string['secondarycolour_desc'] = 'Segunda cor da marca';
+$string['customstylessettings'] = 'Estilos personalizados';
+$string['favicon'] = 'Favicon personalizado';
+$string['favicon_desc'] = 'Deve enviar um ficheiro .ico para usar como favicon.';
+$string['fullwidthpage'] = 'Full width';
+$string['fullwidthpage_desc'] = 'Full width em todas as páginas. Se estiver desselecionado aplica-se o comportamento standard de base.';
+$string['leftcolumn'] = 'Coluna esquerda.';
+$string['leftcolumn_desc'] = 'Conteúdo HTML para ser lido na coluna esquerda do rodapé. Normalmente é uma lista de links, mas pode ser qualquer conteúdo. Tenha em mente apenas que o rodapé está dividido em 3 colunas.';
+$string['centercolumn'] = 'Coluna central.';
+$string['centercolumn_desc'] = 'Conteúdo HTML para ser lido na coluna central do rodapé. Normalmente é uma lista de links, mas pode ser qualquer conteúdo. Tenha em mente apenas que o rodapé está dividido em 3 colunas.';
+$string['backtotopbutton'] = 'Voltar ao topo da página';
+$string['backtotopbutton_desc'] = 'Opção para mostrar ícone para topo da página como nas versões < 4.0 do Boost';
+$string['management'] = 'Gerir';
+$string['searchcourses'] = 'Pesquisar cursos';
+$string['managecourses'] = 'Gerir cursos';
+$string['searchusers'] = 'Pesquisar utilizadores';
+$string['coursecompletion_desc'] = 'Mostrar nível da conclusão do curso do próprio utilizador na página principal do curso';
+$string['frontpagestream'] = 'Opções página inicial';
+$string['catwidgetheading'] = 'Widget de categorias';
+$string['catwidget'] = 'Mostrar widget com categorias';
+$string['catwidget_desc'] = 'Mostrar um layout de categorias de cursos numa grelha responsiva.';
+$string['catwidgetcolumns'] = 'Número de colunas de layout';
+$string['catwidgetcolumns_desc'] = 'Quantas colunas devemos usar na grelha de layout.';
+$string['featuredcategories'] = 'Categorias em destaque';
+$string['featuredcategoriescopy'] = 'A escola de Feitiçaria oferece uma panóplia de temas para escolher. <a href="course/" >Explore-os!</a>';
+$string['maxcatnumber'] = 'Número de categorias a apresentar';
+$string['maxcatnumber_desc'] = 'Número total de categorias que devemos apresentar no widget. A sua pesquisa será limitada a este número.';
+$string['catwidgetimage'] = 'Imagem para cartão de categoria';
+$string['catwidgetimage_desc'] = 'Pode enviar uma imagem padrão para ser apresentada se uma categoria não tiver qualquer imagem provinda das disciplinas. Esta imagem será repetida se existir mais do que uma categoria nestas condições.';
+$string['bunnyfonts'] = 'Tipo de letra';
+$string['bunnyfonts_desc'] = 'Escolha que tipo de letra disponibilizado pelo Bunny Fonts deverá ser usado no site. Este serviço está sediado na UE e não armazena informação sobre os pedidos dos utilizadores. Mais info: https://bunny.net/fonts/';
+$string['externalfonts'] = 'Fontes externas';
+$string['externalfonts_desc'] = 'Ativar serviço externo de tipos de fontes';
+$string['fonts'] = 'Fonts';
+$string['choosencats'] = 'Categorias a apresentar';
+$string['choosencats_desc'] = 'Escolha as categorias que pretende apresentar no widget. Não adicione demasiadas categorias.';
+$string['counthiddencourses'] = 'Incluir disciplinas ocultas';
+$string['counthiddencourses_desc'] = 'Se ativado contabiliza o número total de disciplinas em cada categoria incluindo as ocultas.';
+$string['noresults'] = 'Sem qualquer resultado a apresentar';
+$string['hideprimarynodes'] = 'Esconder opções do menu principal';
+$string['hideprimarynodes_desc'] = 'Pode-se remover cada uma das opções do menú de primeiro nível';
+$string['stream:changesettings'] = 'Permissão para permitir utilizadores que não administradores possam alterar configurações do tema.';
+$string['circlemodicons'] = 'Ícones circulares';
+$string['circlemodicons_desc'] = 'Mostrar círculos como base dos ícones de recursos e atividades na página do curso e no ecrã para adicionar conteúdos.';
+$string['herotitle'] = 'Título do destaque principal';
+$string['herotitle_desc'] = 'Texto a imprimir por cima do destaque principal em cada item. Cabeçalho Heading 1.';
+$string['heromotto'] = 'Frase de destaque ';
+$string['heromotto_desc'] = 'Frase de promoção a mostrar depois do título.';
+$string['herolink'] = 'Endereço da hiperligação';
+$string['herolink_desc'] = 'Endereço para criar o link no botão do destaque principal. Também apresenta um botão call to action.';
+$string['herolinktext'] = 'Todas as disciplinas';
+$string['sliderbutton'] = 'Texto do botãoButton text';
+$string['sliderbutton_desc'] = 'Em função do URL adicionado, pode especificar aqui um texto a apresentar no botão';
+$string['homepageheroimage'] = 'Imagem do destaque principal';
+$string['homepageheroimage_desc'] = 'Imagem a usar como fundo no destaque principal.';
+$string['homepagepromoboximage'] = 'Imagem na caixa de promo';
+$string['homepagepromoboximage_desc'] = 'Imagem a apresentar do lado direito da caixa de promoção.';
+$string['homepagepromoboxwidget'] = 'Widget de promoção';
+$string['homepagepromoboxwidget_desc'] = 'Mostrar uma secção de marketing com uma mensagem, um botão do lado esquerdoe e uma imagem do çado direito.';
+$string['homepagepromoboxtitle'] = 'Título da secção';
+$string['homepagepromoboxtitle_desc'] = 'Texto a usar como cabeçalho para a secção promocional.';
+$string['homepagepromoboxtext'] = 'Corpo de texto';
+$string['homepagepromoboxtext_default'] = 'Uma escola de feiticeiros (também conhecida como escola mágica) é uma instituição educacional que fornece formação mágica a jovens bruxos e feiticeiros. Queres tentar? Regista-te agora e divirte-te!';
+$string['homepagepromoboxtext_desc'] = 'Texto a apresentar do lado esquerdo da secção promocional.';
+$string['homepagepromoboxbutton'] = 'Texto do botão';
+$string['homepagepromoboxbutton_desc'] = 'Texto a apresentar no botão da secção.';
+$string['homepagepromoboxurl'] = 'Link do botão';
+$string['homepagepromoboxurl_desc'] = 'URL a usar como link para o botão. Por omissão remete para a página de registo';
+$string['promoboxwidgetheading'] = 'Widget com caixa de promoção';
+$string['featuredcourseswidgetheading'] = 'Widget de disciplinas em destque';
+$string['featuredcourses'] = 'Disciplinas em destaque';
+$string['featuredcoursescopy'] = 'Com os nossos cursos onlibe pode aprender a qualquer hora, em qualquer sítio ao seu ritmo <a href="/course" >Explore-os!</a>';
+$string['featuredcourseswidget'] = 'Secção de disciplinas em destaque';
+$string['featuredcourseswidget_desc'] = 'Mostra uma secção onde são apresentadas as disciplinas destacadas em formato de cartões, num layout de 3 colunas. Por omissão, são obtidas as disciplinas visíveis que têm data de início no futuro. Check the following settings to check how to tweek that.';
+$string['featuredcoursesshowcat'] = 'Apresentar a categoria';
+$string['featuredcoursesshowcat_desc'] = 'Mostra a categoria da disciplina no topo de cada cartão.';
+$string['featuredcoursesshowcatmax'] = 'Máximo de disciplinas';
+$string['featuredcoursesshowcatmax_desc'] = 'Qual o número máximo que devemos obter a partir da base de dados';
+$string['featuredcoursesfuture'] = 'Mostrar apenas disciplinas futuras';
+$string['featuredcoursesfuture_desc'] = 'Se ativado são apresentadas apenas as disciplinas que se iniciam numa data futura.';
+$string['featuredcoursesfuturerange'] = 'Intervalo de tempo';
+$string['featuredcoursesfuturerange_desc'] = 'O período até ao qual devem ser apresentadas as disciplinas que se iniciem.';
+$string['featuredcoursesrating'] = 'Mostrar pontuação do curso';
+$string['featuredcoursesrating_desc'] = 'Tem de instalar o plugin Course Rating e ativar esta opção de forma a que a pontuação sob forma de estrelas (até 5) seja apresentada em cada cartão.';
+$string['featuredcoursesstartdate'] = 'Mostrar a data de início da disciplina';
+$string['featuredcoursesstartdate_desc'] = 'Se ativado apresenta a data em que começa cada disciplina.';
+$string['availablesoon'] = 'Em breve';
+$string['featuredcoursesshowcatfirst'] = 'Apresentar categoria de primeiro nível';
+$string['featuredcoursesshowcatfirst_desc'] = 'Se ativado, apresentar a categoria de primeiro nível em vez da subcategoria direta a que pertence a disciplina.';
+$string['loginimg'] = 'Imagem da página de autenticação';
+$string['loginimg_desc'] = 'A página de autenticação é apresentada em duas colunas: formulário de autenticação à esquerda e uma imagem à direita. Aqui pode enviar a imagem que pretender.';
+$string['usingstreamtheme'] = 'a usar o tema Stream';
+$string['coursecardimage'] = 'Imagem para cartão da disciplina';
+$string['coursecardimage_desc'] = 'Imagem a usar em cartões de disciplinas que não tenham imagem definida. Se isto se verificar e se esta configuração também estiver em falta, a imagem a usar será aquela definida na pasta default do tema.';
+$string['modiconcoloradministration'] = 'Cor do ícone das atividades de Administração';
+$string['modiconcoloradministration_desc'] = 'Escolha a cor que pretende usar como fundo dos ícones das atividades pertencentes à categoria de \'Administração\'';
+$string['modiconcolorassessment'] = 'Cor do ícone das atividades de Avaliação';
+$string['modiconcolorassessment_desc'] = 'Escolha a cor que pretende usar como fundo dos ícones das atividades pertencentes à categoria de \'Avaliação\'';
+$string['modiconcolorcollaboration'] = 'Cor do ícone das atividades de Colaboração';
+$string['modiconcolorcollaboration_desc'] = 'Escolha a cor que pretende usar como fundo dos ícones das atividades pertencentes à categoria de \'Colaboração\'';
+$string['modiconcolorcommunication'] = 'Cor do ícone das atividades de Comunicação';
+$string['modiconcolorcommunication_desc'] = 'Escolha a cor que pretende usar como fundo dos ícones das atividades pertencentes à categoria de \'Comunicação\'';
+$string['modiconcolorcontent'] = 'Cor do ícone das atividades de Conteúdo';
+$string['modiconcolorcontent_desc'] = 'Escolha a cor que pretende usar como fundo dos ícones das atividades pertencentes à categoria de  \'Conteúdo\ (ou recursos)\'';
+$string['modiconcolorinterface'] = 'Cor do ícone das atividades de Interface';
+$string['modiconcolorinterface_desc'] = 'Escolha a cor que pretende usar como fundo dos ícones das atividades pertencentes à categoria de \'Interface\'';
+$string['modiconoutline'] = 'Contorno dos ícones a preto';
+$string['modiconoutline_desc'] = 'Se ativado,os ícones das atividades usarão como contorno uma cor preta forte.';
+$string['prevactivity'] = 'Atividade anterior';
+$string['nextactivity'] = 'Atividade seguinte';
+$string['slidersettingheading'] = 'Conteúdos do Slider';
+$string['slidestotal'] = 'Número de itens';
+$string['slidestotal_desc'] = 'Escolha aqui o número de itens que quer apresentar no Slider: Tem de gravar a alteração antes de poder inserir a informação.';
+$string['prev'] = 'Anterior';
+$string['next'] = 'Seguinte';
+$string['jumptoslide'] = 'Ir para slide';